Story/77316

+++ src/SME.SGP.Infra/Dtos/Ocorrencias/Detalhes/OcorrenciaDto.cs
@@ -12,8 +12,8 @@ namespace SME.SGP.Infra
         public string OcorrenciaTipoId { get; set; }
         public bool ConsideraHistorico { get; set; }
         public int AnoLetivo { get; set; }
-        public int DreId { get; set; }
-        public int UeId { get; set; }
+        public long DreId { get; set; }
+        public long UeId { get; set; }
         public int Modalidade { get; set; }
         public int Semestre { get; set; }
         public long? TurmaId { get; set; }
